+++ RELEASE-NOTES.md
@@ -23,9 +23,19 @@
 * Cleaning up the output directory after coping files to the stage for Streamlit, Notebook, SPCS Service and project.
 
 ## Fixes and improvements
+
+
+# v3.7.1
+
+## Deprecations
+
+## New additions
+
+## Fixes and improvements
 * Fix certificate connection issues.
 * Fix `snow spcs image-registry login` slow query problem.
 * Added interactive mode for `snow sql`
+
 
 
 # v3.7.0